Update .editorconfig
Updates the editorconfig to match the style new ktp additions should
follow. This commit should *not* be understood as a general reformatting
of the entire spigot/paper codebase, as such as patch would be
unthinkably large for no apperant reason, but rather as a guidline for
new additions in ktp patches.