Add test plugin layout to ease testing api changes (#3)
To ease testing new additions to the api, this commit adds a test plugin
that may be used to develop locally. The test plugin will automatically
be added to any server ran through 'runDev' or the likes if enabled.

To ensure no changes to the test plugin are commited to the repository,
it has to be excluded from git future tracking through the following
bash commands:
 - echo "test-plugin" >> .git/info/exclude
 - find test-plugin -type f | xargs git update-index --assume-unchanged
And uncommenting the include line in 'test-plugin.settings.gradle.kt'.