Expose low level interceptor for vehicle steering
To allow plugins to run their own steering logic when a player is
steering an entity that, by vanilla's design choices, does not react to
vehicle movement packets, this patch introduces new API plugins can use
to handle these packets.

To keep the API as fast and lightweight as possible, the spigot event
bus is not used to allow mutation of these values as creation of an
event and the submission to the handler lists is a lot of overhead for
something called as often as this.

Instead, plugins may register their own custom
PlayerVehicleSteerInterceptor implementation inside the server, which
will be called with the unwrapped values of the packet to allow plugins
to mutate the values send.