Allow events to run on either main or off thread
To allow events that do not know what thread they will occure on to
still be submitted to the event bus, this commit further optimizes the
spigot event bus by representing the async state as a tri state,
allowing events to NOT_SET them.

The simple plugin manager will skip any form of thread validation if an
event may be called on either main or off thread.