Add manual totem of undying resurrection API
This patch introduces new API methods to the living entity that allow
plugins to manually trigger a totem of undying resurrection on any given
living entity.

This is specifically useful for plugins that use the #setHealth methods
to manipulate the entities health.

The API is a full representation of the server internal and fully
simulates a resurrection including bukkit event calls and potion
effects.