Add #getConfigurationSectionList to yaml API
The yaml parser does not consider yaml sections that are defined as part
of a list, configuration sections but rather simple maps. This makes
interaction with theses sections rather difficult.

To ease interaction, this commit creates a new getter in the
configuration section interface that returns a list of configuration
sections by actively converting the parsed maps into memory sections.