Expand de/serializeEntity API for adventure-nbt
With the addition of adventure-nbt, an API level representation of nbt
compounds are available to plugins.
While they are not widely used, entity storage is a perfect place for
them.

This commit adds a new de- and serialization method for entities that
returns the adventure-nbt representation of the compound tag.