Add item stack to SNBT serialiser
Adds a new method to the UnsafeValues interface that allows developers
to convert an API item stack into the SNBT required to create it.
This method allows for a complete string representation that is human
readable compared to the existing ItemStack#toString() method that does
only expose metadata known to the ItemMeta.