Show server as enforceSecureProfile behind proxy
When a server is behind a proxy server, its online-mode setting locally
is set to false as authentication is left to the proxy server.
While modern forwarding under velocity allows for an online-moder
server, bungee does not.

To prevent the popup under a bungee system that requires a full
system-message chat setup, this patch defines a server as
enforceSecureProfile if it is configured to be in a proxy-online mode.