Add considerInvalidMessageSignatureAsUnsigned
The considerInvalidMessageSignatureAsUnsigned option added by this patch
allows the server to consider any player chat message that was sent that
contains an invalid or missing signature to simply interpet as a
unsigned message.

This behaviour may break basically everything unless a respective plugin
is in place to also map all messages to system messages, which is the
case in the environment KTP is meant to be run.

This patch is certainly not a final fix for the current problem facing
the production environment ktp is deployed in, but works as a quick and
dirty fix for the issue.