Rework plugin translatables to translatable
The PluginTranslatableComponent framework, while effective for packet
based messages, has, as obvious during its conception, proven too
difficult to nicely adapt into the eco system as usage of these third
party components causes untrackable errors when applied to anything not
purely packet based.

This commit aims to rework, and hence also break, the API that is not
functional enough for the needs of KTP and move it to a new framework
that uses plain TranslatableComponent instances but makes use of
adventures new ability to run custom formatting logic in the translator,
which allows the mini message translation to still work.