Optimise PaperSeenByTracker#contains checks
Fundamentally, the seen by tracker works on the assumption that
modifications to its state are rare compared to reads or non-modifying
calls to add/remove.

Similar to the original replacement of the set, this assumption is based
on the continuous calls to #add and #remove each tick for entities that
are or aren't tracking the entity already respectively.
The ReferenceSet implementation for #add and #remove is not free, as
it has to compute the table key, resolve possible collisions *and* then
perform the equality check for the key.

Alternatively, this patch assigns a continuousIdentifier to each
ServerPlayerConnection on the server, which is guaranteed to be at
*most* the number of concurrent players on the server during the JVMs
lifetime.
The identifier is *near* continuous, as players may leave.
Open identifiers from players left are allocated to joining players.
This enables the usage of a BitSet as a mirroring state for the trackers
map "containing" a player.
With player counts below 64 players, this is doable in a single long.

! NOTICE !
The impact of this patch is not *easy* to profile outside of actual
server workload as the add/remove operations are well below a
millisecond and hence are easily poisoned by the required state setup
of the to-be-mutated maps in jmh.
As such, the impact will only be known once the patch has lived long
enough in this fork and ran in production for detailed profiling.